Moved config_system to system.config
