+++ setup.py
@@ -278,8 +278,8 @@ package_data['sas.qtgui'] = ['Calculators/UI/*',
                              ]
 packages.append("sas.qtgui")
 
-package_data["sas.config_system"] = ["*"]
-packages.append("sas.config_system")
+package_data["sas.config"] = ["*"]
+packages.append("sas.config")
 
 required = [
     'bumps>=0.7.5.9', 'periodictable>=1.5.0', 'pyparsing>=2.0.0',
