+++ src/sas/__init__.py
@@ -2,7 +2,7 @@ from sas.sasview import __version__
 
 import os
 import sys
-from sas.config_system import configuration as config
+from sas.system.config import configuration as config
 
 __all__ = ['get_app_dir', 'get_user_dir', 'config']
