--- src/sas/config_system/__init__.py
@@ -1,1 +1,0 @@
-from sas.config_system.util import configuration
