+++ src/sas/system/config/__init__.py
@@ -1,0 +1,1 @@
+from sas.system.config.util import configuration
